moved request object to representations, minor cleanup
